add docs, move stuff to wasm mod
