+++ libsql/Cargo.toml
@@ -106,3 +106,8 @@ cloudflare = [
 [[bench]]
 name = "benchmark"
 harness = false
+
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
